BC BREAK: tasks_lib: update sub_run
Makes it more friendly for usage and error troubleshooting.

BC BREAK: when doing a capture, this newer version will also set "text=True" which means the output will be decoded automatically.  So if you are using `result.stdout.decode(...)` or `result.stderr.decode(...) you will want to remove the decode().

This is only a default.  If you want to keep the current behavior, use `sub_run(..., text=False)`.